+++ packages/config/src/config.ts
@@ -1,4 +1,5 @@
 import { Address, NocturneContractDeployment } from "./deployment";
+import * as sepolia from "../configs/sepolia.json";
 import * as localhost from "../configs/localhost.json";
 import * as exampleNetwork from "../configs/example-network.json";
 import * as fs from "fs";
@@ -107,11 +108,14 @@ export function loadNocturneConfig(
 }
 
 export function loadNocturneConfigBuiltin(name: string): NocturneConfig {
-  if (name == "localhost") {
-    return NocturneConfig.fromObject(localhost as any);
-  } else if (name == "example-network") {
-    return NocturneConfig.fromObject(exampleNetwork as any);
-  } else {
-    throw new Error(`unknown config name: ${name}`);
+  switch (name) {
+    case "sepolia":
+      return NocturneConfig.fromObject(sepolia as any);
+    case "localhost":
+      return NocturneConfig.fromObject(localhost as any);
+    case "example-network":
+      return NocturneConfig.fromObject(exampleNetwork as any);
+    default:
+      throw new Error(`unknown config name: ${name}`);
   }
 }